deploy: Move treasury deploy to deploy_contracts
