+++ deploy/deploy_contracts.ts
@@ -1,7 +1,7 @@
 import {HardhatRuntimeEnvironment} from "hardhat/types"
 import {DeployFunction} from "hardhat-deploy/types"
 import {ethers} from "hardhat"
-import {Contract} from "ethers"
+import {Contract, constants} from "ethers"
 import {DeployResult, Export} from "hardhat-deploy/dist/types"
 import fs from "fs"
 
@@ -11,7 +11,9 @@ import {
     RoundsManager,
     TicketBroker,
     LivepeerToken,
-    Governor
+    Governor,
+    LivepeerGovernor,
+    Treasury
 } from "../typechain"
 
 import ContractDeployer from "../utils/deployer"
@@ -212,6 +214,21 @@ const func: DeployFunction = async function(hre: HardhatRuntimeEnvironment) {
     await (await Governor.stage(transferOwnershipUpdate, 0)).wait()
     await (await Governor.execute(transferOwnershipUpdate)).wait()
 
+    // Treasury (LivepeerGovernor)
+
+    const treasury = await contractDeployer.deployAndRegister({
+        contract: "Treasury",
+        name: "Treasury",
+        args: []
+    })
+
+    const livepeerGovernor = await contractDeployer.deployAndRegister({
+        contract: "LivepeerGovernor",
+        name: "LivepeerGovernor",
+        args: [Controller.address],
+        proxy: true
+    })
+
     // Set BondingManager parameters
     const BondingManager: BondingManager = (await ethers.getContractAt(
         "BondingManager",
@@ -258,6 +275,48 @@ const func: DeployFunction = async function(hre: HardhatRuntimeEnvironment) {
             ).wait()
         }
     }
+
+    // Initialize Treasury and LivepeerGovernor
+
+    const Treasury: Treasury = await ethers.getContractAt(
+        "Treasury",
+        treasury.address
+    )
+
+    await Treasury.initialize(
+        config.treasury.minDelay,
+        [], // governor will be added as a proposer later
+        [constants.AddressZero], // let anyone execute proposals
+        deployer // temporary admin role for deployer
+    ).then(tx => tx.wait())
+
+    const LivepeerGovernor: LivepeerGovernor = await ethers.getContractAt(
+        "LivepeerGovernor",
+        livepeerGovernor.address
+    )
+
+    await LivepeerGovernor.initialize(
+        config.livepeerGovernor.initialVotingDelay,
+        config.livepeerGovernor.initialVotingPeriod,
+        config.livepeerGovernor.initialProposalThreshold,
+        config.livepeerGovernor.initialQuorum,
+        config.livepeerGovernor.quota
+    ).then(tx => tx.wait())
+
+    // Now grant proposer and executor roles to governor and renounce deployer admin role
+    const roles = {
+        proposer: await Treasury.PROPOSER_ROLE(),
+        canceller: await Treasury.CANCELLER_ROLE(),
+        executor: await Treasury.EXECUTOR_ROLE(),
+        admin: await Treasury.TIMELOCK_ADMIN_ROLE()
+    }
+    for (const role of [roles.proposer, roles.canceller, roles.executor]) {
+        await Treasury.grantRole(role, LivepeerGovernor.address).then(tx =>
+            tx.wait()
+        )
+    }
+
+    await Treasury.renounceRole(roles.admin, deployer).then(tx => tx.wait())
 
     // Set TicketBroker parameters
     const Broker: TicketBroker = (await ethers.getContractAt(
